session.go: ignore EmptyAtom tokens
For some drives we see that the response contains some EmptyTokens
The parsing done today expect status in specific location from
the end of reply and these tokens parse reply as malformed.

This fix will trim EmptyAtoms if exist at the end of response.

according to 3.2.2.3.1.5 Empty Atom:

The Empty atom does not encode values.
The Empty atom allows other values in the data subpacket
contained in that part of the stream to be aligned with multi-byte
boundaries for efficiency. It also allows areas of a fixed buffer to
be filled with a value that is able to be safely ignored.